\name{PDSR-package}
\alias{PDSR-package}
\alias{PDSR}
\title{
An R package for NASA's Planetary Data System
}
\description{
PDSR use NASA's FTP service to smoothly access PDS. Through metadata
file parsing, it helps the data-scientist downloading and describing
data without using the web-browser.

This package was designed during the 2013 version of the NASA's 
International SpaceApps Challenge

Please visit the project home for more information:
\url{https://github.org/gvegayon/PDSR}.

or its SpaceApps repo

\url{https://github.org/spaceappcl/team20}.
}
\details{
  \tabular{ll}{
    Package: \tab PDSR\cr
    Type: \tab Package\cr
    Version: \tab 0.13.04\cr
    Date: \tab 2013-04-21\cr
    License: \tab GPL version 2 or later\cr
  }
}
\author{
George G. Vega \email{g.vegayon@gmail.com},

Joshua B. Kunst \email{jbkunst@gmail.com}
}
