\name{getColnames}
\alias{getColnames}
\title{
Reaches colnames of every table within a PDS object
}
\description{
Reaches colnames of every table within a PDS object
}
\usage{
getColnames(x)
}
\arguments{
  \item{x}{A PDSR class object}
}
\details{
this and that
}
\value{
Nested list containing tables and their colnames
}
% \references{}
\author{
George
}
%\note{%%  ~~further notes~~}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{\code{\link{help}}}
% \examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ IO }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
